feat(reservedetails): add orderId to reserve details interface and model
This commit introduces the `orderId` field to the `IReserveDetails` interface and the corresponding schema in the `ReserveDetailsModel`. The `orderId` is generated as a random 6-digit number during the creation of reserve details. This change is necessary to uniquely identify each reservation for better tracking and management.

Additionally, unnecessary console logs were removed from the controller, and the `localField` in the cars service was updated from `userId` to `agencyId` to align with the correct data structure.